[feat] 충전 로직 위한 디렉토리
