+++ src/main/java/kr/hhplus/be/server/application/point/ChargePointCommand.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.application.point;
+
+public class ChargePointCommand {
+
+}
