+++ src/main/java/kr/hhplus/be/server/application/point/ChargePointResult.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.application.point;
+
+public class ChargePointResult {
+
+}
