+++ src/main/java/kr/hhplus/be/server/application/point/UserPointFacade.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.application.point;
+
+public class UserPointFacade {
+
+}
