--- src/main/java/kr/hhplus/be/server/coupon/dto/CouponListResponse.java
@@ -1,9 +1,0 @@
-package kr.hhplus.be.server.coupon.dto;
-
-
-import java.util.List;
-
-public record CouponListResponse(
-        Long userId,
-        List<CouponResponse> coupons
-) { }
