--- src/main/java/kr/hhplus/be/server/coupon/dto/CouponResponse.java
@@ -1,12 +1,0 @@
-package kr.hhplus.be.server.coupon.dto;
-
-import java.time.LocalDate;
-
-public record CouponResponse(
-        Long id,
-        String title,
-        String discountType, // "RATE" or "AMOUNT"
-        Integer discountValue,
-        LocalDate startDate,
-        LocalDate endDate
-) { }
