+++ src/main/java/kr/hhplus/be/server/domain/BusinessException.java
@@ -1,0 +1,15 @@
+package kr.hhplus.be.server.domain;
+
+public class BusinessException extends RuntimeException {
+    private final int status;
+
+    public BusinessException(int status, String message) {
+        super(message);
+        this.status = status;
+    }
+
+    public int getStatus() {
+        return status;
+    }
+}
+
