+++ src/main/java/kr/hhplus/be/server/domain/point/UserPointRepository.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.domain.point;
+
+public class UserPointRepository {
+
+}
