+++ src/main/java/kr/hhplus/be/server/domain/point/UserPointService.java
@@ -1,0 +1,13 @@
+package kr.hhplus.be.server.domain.point;
+
+public class UserPointService {
+
+    public void charge(UserPoint userPoint, Long amount){
+        userPoint.charge(amount);
+    }
+
+    public  void validateCharfe(UserPoint userPoint, Long amount){
+
+    }
+
+}
