+++ src/main/java/kr/hhplus/be/server/infrastructure/point/InMemoryUserPointRepository.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.infrastructure.point;
+
+public class InMemoryUserPointRepository {
+
+}
