+++ src/main/java/kr/hhplus/be/server/interfaces/GlobalExceptionHandler.java
@@ -1,0 +1,34 @@
+package kr.hhplus.be.server.interfaces;
+
+
+import kr.hhplus.be.server.domain.BusinessException;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler(BusinessException.class)
+    public ResponseEntity<ErrorResponse> handleBusiness(BusinessException e) {
+        return ResponseEntity.status(e.getStatus())
+                .body(new ErrorResponse(e.getMessage()));
+    }
+
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<ErrorResponse> handleAll(Exception e) {
+        return ResponseEntity.status(500)
+                .body(new ErrorResponse("서버 내부 오류가 발생했습니다."));
+    }
+
+    public static class ErrorResponse {
+        private final String message;
+
+        public ErrorResponse(String message) {
+            this.message = message;
+        }
+
+        public String getMessage() {
+            return message;
+        }
+    }
+}
