+++ src/main/java/kr/hhplus/be/server/interfaces/point/ChargePointRequest.java
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.interfaces.point;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+public class ChargePointRequest {
+
+}
