+++ src/main/java/kr/hhplus/be/server/interfaces/point/ChargePointResponse.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.interfaces.point;
+
+public class ChargePointResponse {
+
+}
