+++ src/main/java/kr/hhplus/be/server/interfaces/point/PointController.java
@@ -1,0 +1,10 @@
+package kr.hhplus.be.server.interfaces.point;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/v1/points")
+public class PointController {
+
+}
