+++ src/main/java/kr/hhplus/be/server/lagacy/coupon/dto/CouponIssueRequest.java
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.lagacy.coupon.dto;
+
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotNull;
+
+public record CouponIssueRequest(
+        @NotNull @Min(1) Long userId
+) { }
