+++ src/main/java/kr/hhplus/be/server/lagacy/coupon/dto/CouponListResponse.java
@@ -1,0 +1,9 @@
+package kr.hhplus.be.server.lagacy.coupon.dto;
+
+
+import java.util.List;
+
+public record CouponListResponse(
+        Long userId,
+        List<CouponResponse> coupons
+) { }
