+++ src/main/java/kr/hhplus/be/server/lagacy/order/dto/OrderCreateResponse.java
@@ -1,0 +1,7 @@
+package kr.hhplus.be.server.lagacy.order.dto;
+
+import java.util.List;
+
+public record OrderCreateResponse(
+        Long orderId
+) { }
