+++ src/main/java/kr/hhplus/be/server/lagacy/order/dto/OrderItemRequest.java
@@ -1,0 +1,9 @@
+package kr.hhplus.be.server.lagacy.order.dto;
+
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotNull;
+
+public record OrderItemRequest(
+        @NotNull @Min(1) Long productId,
+        @NotNull @Min(1) Integer quantity
+) { }
