+++ src/main/java/kr/hhplus/be/server/lagacy/point/dto/PointChargeResponse.java
@@ -1,0 +1,6 @@
+package kr.hhplus.be.server.lagacy.point.dto;
+
+public record PointChargeResponse(
+        Long userId,
+        Long balance
+) { }
