+++ src/main/java/kr/hhplus/be/server/lagacy/product/dto/BestProductListResponse.java
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.lagacy.product.dto;
+
+
+import java.util.List;
+
+public record BestProductListResponse(
+        List<BestProductResponse> products
+) { }
