+++ src/main/java/kr/hhplus/be/server/lagacy/product/dto/BestProductResponse.java
@@ -1,0 +1,9 @@
+package kr.hhplus.be.server.lagacy.product.dto;
+
+public record BestProductResponse(
+        Long id,
+        String name,
+        Long price,
+        Integer sales,
+        Integer stock
+) { }
