+++ src/main/java/kr/hhplus/be/server/lagacy/product/dto/ProductListResponse.java
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.lagacy.product.dto;
+
+
+import java.util.List;
+
+public record ProductListResponse(
+        List<ProductResponse> products
+) { }
