+++ src/main/java/kr/hhplus/be/server/lagacy/product/dto/ProductResponse.java
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.lagacy.product.dto;
+
+public record ProductResponse(
+        Long id,
+        String name,
+        Long price,
+        Integer stock
+) { }
