--- src/main/java/kr/hhplus/be/server/order/dto/OrderCreateResponse.java
@@ -1,7 +1,0 @@
-package kr.hhplus.be.server.order.dto;
-
-import java.util.List;
-
-public record OrderCreateResponse(
-        Long orderId
-) { }
