--- src/main/java/kr/hhplus/be/server/order/dto/OrderItemRequest.java
@@ -1,9 +1,0 @@
-package kr.hhplus.be.server.order.dto;
-
-import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotNull;
-
-public record OrderItemRequest(
-        @NotNull @Min(1) Long productId,
-        @NotNull @Min(1) Integer quantity
-) { }
