--- src/main/java/kr/hhplus/be/server/point/dto/PointChargeRequest.java
@@ -1,10 +1,0 @@
-package kr.hhplus.be.server.point.dto;
-
-import jakarta.validation.constraints.Max;
-import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotNull;
-
-public record PointChargeRequest(
-        @NotNull @Min(1) Long userId,
-        @NotNull @Min(1) @Max(1_000_000) Long chargeAmount
-) { }
