--- src/main/java/kr/hhplus/be/server/point/dto/PointChargeResponse.java
@@ -1,6 +1,0 @@
-package kr.hhplus.be.server.point.dto;
-
-public record PointChargeResponse(
-        Long userId,
-        Long balance
-) { }
