--- src/main/java/kr/hhplus/be/server/point/dto/PointUseRequest.java
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.point.dto;
-
-import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotNull;
-
-public record PointUseRequest(
-        @NotNull @Min(1) Long orderId
-) { }
