--- src/main/java/kr/hhplus/be/server/product/dto/BestProductListResponse.java
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.product.dto;
-
-
-import java.util.List;
-
-public record BestProductListResponse(
-        List<BestProductResponse> products
-) { }
