--- src/main/java/kr/hhplus/be/server/product/dto/BestProductResponse.java
@@ -1,9 +1,0 @@
-package kr.hhplus.be.server.product.dto;
-
-public record BestProductResponse(
-        Long id,
-        String name,
-        Long price,
-        Integer sales,
-        Integer stock
-) { }
