--- src/main/java/kr/hhplus/be/server/product/dto/ProductListResponse.java
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.product.dto;
-
-
-import java.util.List;
-
-public record ProductListResponse(
-        List<ProductResponse> products
-) { }
