--- src/main/java/kr/hhplus/be/server/product/dto/ProductResponse.java
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.product.dto;
-
-public record ProductResponse(
-        Long id,
-        String name,
-        Long price,
-        Integer stock
-) { }
